Comment in OverallUILayout and Add Menu UI layer
Useful comment in W_OverallUILayout to explain how the stacks work and what they're intended for. Added Menu UI layer to match Lyra's OverallUILayout stacks.

That layer might be useful for settings menu, I almost added the Pause menu to it but I think it still counts as a game menu.